requirements.txt
numba==0.64.0
numpy==2.4.3
pandas>=2.0.0,<3.0.0
plotly==6.5.0
protobuf>=3.20,<7
PyYAML==6.0.3
streamlit==1.55.0
tqdm==4.67.3
tushare==1.4.25
akshare>=1.11.0
requests>=2.31.0
openai>=1.0.0
